perf: server separate from routes
